+++ .gitignore
@@ -51,7 +51,7 @@ coverage.xml
 .pytest_cache/
 cover/
 
-api/common/.env
+config/.env
 
 # Translations
 *.mo
